moved eCommerce specific code from BaseSpider
