+++ setup.cfg
@@ -37,4 +37,4 @@ per-file-ignores =
 
     # E731: Ignore "do not assign a lambda expression, use a def" since
     # we're using quick shortcuts for the tests
-    tests/test_base.py:E731
+    tests/test_ecommerce.py:E731
